Butterfly Hybrid Detection/requirements.txt
torch==2.3.0
torchvision==0.18.0
transformers==4.30.0
scikit-learn==1.6.1
opencv-python-headless==4.8.0.76
pandas==2.1.1
Pillow==10.0.0
numpy==1.24.0
